can declare mod view

+++ broccoli_server/mod_view/mod_view_store.py
@@ -1,6 +1,7 @@
 import pymongo
+from .logging import logger
 from broccoli_server.mod_view.mod_view_query import ModViewQuery
-from typing import List, Tuple
+from typing import List, Tuple, Optional
 
 
 class ModViewStore(object):
@@ -8,8 +9,13 @@ class ModViewStore(object):
         self.client = pymongo.MongoClient(connection_string)
         self.db = self.client[db]
         self.collection = self.db["boards"]
+        self.declared_mod_views = []  # type: List[Tuple[str, ModViewQuery]]
 
     def upsert(self, mod_view_id: str, mod_view_query: ModViewQuery):
+        if self.declared_mod_views:
+            logger.warn("Cannot modify persisted mod views if mod views are declared")
+            return
+
         existing_mod_views = []
         # todo: dup with get_all
         for d in self.collection.find():
@@ -38,6 +44,9 @@ class ModViewStore(object):
         })
 
     def get_all(self) -> List[Tuple[str, ModViewQuery]]:
+        if self.declared_mod_views:
+            return self.declared_mod_views
+
         existing_mod_views = []
         for d in self.collection.find().sort("position", pymongo.ASCENDING):
             existing_mod_views.append(
@@ -45,11 +54,21 @@ class ModViewStore(object):
             )
         return existing_mod_views
 
-    def get(self, mod_view_id: str) -> ModViewQuery:
+    def get(self, mod_view_id: str) -> Optional[ModViewQuery]:
+        if self.declared_mod_views:
+            for _id, q in self.declared_mod_views:
+                if mod_view_id == _id:
+                    return q
+            return None
+
         doc = self.collection.find_one({"board_id": mod_view_id})
         return ModViewQuery(doc["board_query"])
 
     def swap(self, mod_view_id: str, another_mod_view_id: str):
+        if self.declared_mod_views:
+            logger.warn("Cannot modify persisted mod views if mod views are declared")
+            return
+
         # todo: find one dups with get()
         mod_view_position = self.collection.find_one({"board_id": mod_view_id})["position"]
         another_mod_view_position = self.collection.find_one({"board_id": another_mod_view_id})["position"]
@@ -63,6 +82,10 @@ class ModViewStore(object):
         )
 
     def remove(self, mod_view_id: str):
+        if self.declared_mod_views:
+            logger.warn("Cannot modify persisted mod views if mod views are declared")
+            return
+
         # todo: shred positions afterwards
         self.collection.delete_one({"board_id": mod_view_id})
 
@@ -71,3 +94,6 @@ class ModViewStore(object):
             print(f"Going to remove {self.collection.count_documents({})} documents")
         else:
             self.collection.delete_many({})
+
+    def declare_mod_views(self, mod_views: List[Tuple[str, ModViewQuery]]):
+        self.declared_mod_views = mod_views
